requirements.txt
django>=3.0.7
pycodestyle==2.6.0
django-crispy-forms
slackclient
